requirements.txt
gitpython>=3.1.31
graphviz>=0.20.1
openai>=0.27.0
rich>=13.3.2
click>=8.1.7
pytest>=7.0.0
